[services] Identity - Adding the generated protobuf files
Summary:
We have updated the protobuf file for the Identity service but forgot to include the generated ones.
This diff fixes the issue where on every native build the updated headers are generated and are different from what we have in a repo.

Test Plan: Services and native build succeeded.

Reviewers: jon, varun

Reviewed By: varun

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5205