[services] Tunnelbroker - Update sending of the `newNotifyTokenRequired` in `Get` stream request
Summary:
This diff introduces sending of the `newNotifyTokenRequired` to the client in `Get` stream request.
When the device has an invalid or expired notification token it's empty in the database and we are requesting a new token from the client using the `newNotifyTokenRequired` response.
The client gets back with the new token in `Get` stream request when starts listening for new messages (getting messages) from the Tunnelbroker (ref to D5206).

This is a part of the stack.
This approach was used keeping in mind the transition to the bidirectional stream soon.

Related Linear task: [[ https://linear.app/comm/issue/ENG-1782/generate-notice-for-client-for-invalid-device-tokens-in-tunnelbroker | ENG-1782 ]] and [[ https://linear.app/comm/issue/ENG-1816/update-tunnelbroker-api-to-support-sending-notices-to-clients-for | ENG-1816 ]]

Test Plan: Successfully built the service and passed unit and integration tests.

Reviewers: jon, tomek

Reviewed By: jon, tomek

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5207