[lib] Implement `messageContentForServerDB` for `messageTypes.MULTIMEDIA`
Summary:
The `messageContentForServerDB(...)` function determines what gets put in the `content` column of the `messages` table.

The current `content` column for media messages is a "stringified" list of `uploadID`s. However, the content column is never used after it's been stored in the DB. Rather, messages are joined with corresponding uploads via the `container` column of the `uploads` table. Therefore this change isn't "destructive."

---

Depends on D5195

Test Plan:
Rows in the `messages` and `uploads` column appear as expected:
{F175860}

{F175861}

Reviewers: abosh, tomek, marcin

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5197