[CI] Bump GH Actions `ios_ci` job to `macos-12`
Summary:
Brings "iOS Build CI" job consistent with "iOS Release" and "iOS XCTest" jobs. Not sure why they were out of sync to begin with... I think it has to do with GitHub lagging behind LTS with the `macos-latest` and `ubuntu-latest` agent tags.

This diff is part of trying to fix the GH Actions iOS CI. Not sure that this diff will make a difference, but the iOS XCTest workflow builds the app successfully so there's a chance?

Test Plan: NA, should be a noop. Land this diff and see what happens.

Reviewers: varun, tomek, marcin

Reviewed By: varun

Subscribers: ashoat, abosh

Differential Revision: https://phab.comm.dev/D5214