[services] Tunnelbroker - Add test for `updateSessionItemDeviceToken` database method
Summary:
This diff introduces a test for the `updateSessionItemDeviceToken` method from D5179.

In this test, we will create a `SessionItem` entity and push it to the DynamoDB table. We are checking the default value and updating it using the `updateSessionItemDeviceToken` method and checks the value to become an updated one.

Related Linear task: [[ https://linear.app/comm/issue/ENG-1782 | ENG-1782 ]]

Test Plan:
1. Successfully built.
2. Test in D5184 is passed.

Reviewers: jon, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D5184