[native] Display thumbnail in `Multimedia` component for videos
Summary:
Now that we have the persistence and retrieval of video thumbnails figured out, we need to actually display them in the `native` app.

In this diff we just retrieve the `thumbnailURI` instead of `uri` from `mediaInfo` for media of type `video`.

This enables a minimally viable end-to-end video experience.

NOTE: Still need to take a closer look at the `reportURIDisplayed(...)` logic and understand what's going on there better before landing this diff.

Test Plan:
See video:
{F178235}

Reviewers: tomek, varun, abosh, marcin

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5216