[flow] Add `(killall flow || pkill flow || true)` to `cleaninstall`
Summary:
When new JS dependencies are added, the `flow` server(s) need to be restarted to avoid inaccurate errors. In the past it's been on the developer to know this and run `killall flow && flow` (or the equivalent) to get their dev environment in working condition again. In this diff we just add that to the beginning of `cleaninstall` to take care of that for us.

`killall` is for macOS/Darwin (though they do have bsd pkill as well)
`pkill` is for Ubuntu/Debian/etc
`true` is for the case where there's no `flow` server running (eg very first run of `yarn cleaninstall`)

Test Plan: Ran `yarn cleaninstall` locally.. the pre-review CI will implicitly test things on Ubuntu/Debian

Reviewers: tomek, varun, marcin, abosh, ginsu, rohan

Reviewed By: varun

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5223