[services] Tunnelbroker - Update device online status in `Get` handler
Summary:
This diff introduces an update of the device's online state in the database when the ping message or the payload message write was unsuccessful.

This is a part of the stack.

Related Linear task: [[ https://linear.app/comm/issue/ENG-1766/make-sure-the-device-is-offline-to-send-a-notification-to-wake-it-up | ENG-1766 ]]

Test Plan: Successfully built the server and the follow-up diffs in the stack.

Reviewers: jon, tomek, karol

Reviewed By: jon, tomek

Subscribers: ashoat, tomek, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D5175