[lib/native] refactor `LogInActionSource` type to use object pattern
Summary:
This diff refactors previous union type `LoginActionSource` and scattered constants with actions to frequently used pattern with the frozen object to maintain a single source of truth for actions.
Other than that, it'll make validation of this field easier while passing it to the keyserver (introduced in next diffs).

Test Plan: Run type checks to make sure ale usages were properly updated.

Reviewers: tomek, atul, abosh

Reviewed By: tomek

Subscribers: ashoat, atul, abosh

Differential Revision: https://phab.comm.dev/D5226