[keyserver] allow `source` field in log_in endpoint validation
Summary:
We will need to pass the `source` field to the keyserver to distinguish between different log-in scenarios ([[ https://linear.app/comm/issue/ENG-1867/pass-loginactionsource-to-keyserver-when-logging-in | task ]]).
This diff adds this field to the `strict` interface but allows an `undefined` source (which will happens on logging from the login form).
A proper type for login from form will be added later on and this field will become mandatory.

Test Plan:
1. Modify [[ https://github.com/CommE2E/comm/blob/0c29c477a3e79efd81b7aadfc7c483883d89170a/lib/actions/user-actions.js#L132 | the code ]] to pass the `source` field.
2. Check if validation pass for both `undefined` (e.g. login from form) and defined source (e.g. cookie expired).

Reviewers: tomek, atul, abosh

Reviewed By: tomek

Subscribers: ashoat, atul, abosh

Differential Revision: https://phab.comm.dev/D5227