[web] Disable 'Delete' button for threads when the password input is empty
Summary:
Added a check in the `disabled` property of the delete button to
check if there is an password entered into the input field - if not,
disable the button.

Test Plan:
Tested on web and it works as expected. Handled cases where the
modal is opened for the first time, and also when a password is deleted.
The button remains disabled whenever the input is empty.

{F184802}

Reviewers: atul, jakub, ginsu

Reviewed By: atul, ginsu

Subscribers: ashoat, tomek, abosh

Differential Revision: https://phab.comm.dev/D5249