[web/native/lib] make `source` param required on log in
Summary:
Make `source` param required while calling the server endpoint. ([[ https://linear.app/comm/issue/ENG-1870/make-source-a-required-parameter-to-login-function-in-lib | task ]])

Note: there are still some places where `source` is optional but it is used in a different context - for now we care only about making it required while connecting with the keyserver, but if it is signification can be refactored in the future to make `source` required in remaining places.

Test Plan:
1. Running `flow check`.
2. Checking all callers of `resolveInvalidatedCookie` - it's called only by `fetchNewCookieFromNativeCredentials` and `source` was [[ https://github.com/CommE2E/comm/blob/013cf5a51ca823409699ad22dc7e9b8914d59dab/lib/utils/action-utils.js#L181 | required there previously ]] and everywhere `fetchNewCookieFromNativeCredentials` is called `source` is passed explicitly from `loginActionSources`.

Reviewers: tomek, atul, ashoat

Reviewed By: ashoat

Subscribers: ashoat, atul, abosh

Differential Revision: https://phab.comm.dev/D5259