[services] Tunnelbroker - Renaming Rust library directory from `rust-notifications` to `rust-lib`
Summary:
This diff renaming the Rust code directory from `rust-notifications` to `rust-lib` name, because we are going to use a single Rust library with the submodules instead of building separate Rust libraries.

This is a part of the stack.
The full context is available in the Linear task.

Related Linear task: [[ https://linear.app/comm/issue/ENG-1895/change-to-use-a-single-rust-code-library-to-c-code-instead-of-a-few | ENG-1895 ]]

Test Plan: Service is successfully built and the Rust function is successfully called from the C++ side.

Reviewers: jon, michal, marcin, varun, tomek

Reviewed By: jon, marcin, tomek

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5261