[services] Tunnelbroker - Adding of Tonic crate to cargo dependencies
Summary:
This diff introduces adding the [[ https://github.com/hyperium/tonic | Tonic ]] to cargo dependencies for the Rust library in Tunnelbroker.
The purpose of adding Tonic is to use it for gRPC communication.

Related Linear task: [[ https://linear.app/comm/issue/ENG-1072/use-rust-tonic-crate-in-tunnelbroker-to-implement-messagesstream | ENG-1072 ]]

Test Plan: No test plan, this is just a dependency.

Reviewers: jon, tomek, varun

Reviewed By: jon, tomek

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5265