Split createMessageReply to createMessageReply and createMessageQuote
Summary:
In message reporting we need to create a quote that will be placed at the end of a message and as such doesn't need newlines at the end.
Related discussion in D4941.

Test Plan: See that replying to a message still works, even for messages with multiple newlines in a row.

Reviewers: tomek, michal, jakub, ashoat

Reviewed By: tomek, ashoat

Subscribers: atul, abosh

Differential Revision: https://phab.comm.dev/D5212