[web] modified InlineSidebar container width to fix horizontal scrolling in ChatMessageList
Summary:
modified `InlineSidebar` container with `margin-left` and `margin-right`  properties to get the correct container width and get rid of horizontal scrolling in `ChatMesssageList`

---

Linear Task: [[ https://linear.app/comm/issue/ENG-1911/horizontal-scrollbar-visible-in-chat-on-web | ENG-1911 ]]

Test Plan:
Please view the screenshots to see how the `InlineSidebar` width was changed. Please note that the red border is just there to highlight the change and is not part of the changes. Additionally, to test the no scrolling of the `ChatMessageList` go to a Chat that has a thread in it. Try scrolling left and right, and you should not be able to do it, which is the expected behavior.

Before:
{F187428}

{F187424}
{F187425}

After:
{F187426}
{F187427}

Reviewers: atul, abosh, rohan, michal

Reviewed By: atul, rohan

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5271