[web] Relocate the icon in ThreadSettingsDeleteTab to be in-line with the warning text
Summary:
Moved the i icon to be in-line with the warning text when
attempting to delete a chat. This was done by moving the SWMansionIcon
element to be inside the paragraph element.

Addresses: https://linear.app/comm/issue/ENG-1921/relocate-the-icon-in-threadsettingsdeletetab-to-be-in-line-with-the

Test Plan:
Confirmed visually that this change works as expected, and will
attach before/after photos so reviewers can see.

Before:
{F187504}

After:
{F187505}

Reviewers: atul, ginsu, ashoat

Reviewed By: atul, ginsu, ashoat

Subscribers: ashoat, tomek, abosh

Differential Revision: https://phab.comm.dev/D5273