[web] Notifications dialog - add text
Summary:
[ENG-1729](https://linear.app/comm/issue/ENG-1729/update-notifications-dialog-to-handle-threads-correctly)
Adds additional explanation text. For sidebars in the Focused tab, it's displayed under the button and for sidebars in the Background tab it's displayed instead of previous content.

Test Plan: Check that nothing changed for non-sidebars. Check that for each combination of *Parent is in Focused/Background tab* and *can/can't be promoted* for sidebars is displayed correctly.

Reviewers: tomek, atul, abosh

Reviewed By: tomek, atul

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5281