Implement JNI Helper for GlobalDBSingleton
Summary: This differential exposes GlobalDbSingleton API to Java native code through JNI. I understand that this differential is quite big, but all that actually matters is the .cpp file. The rest is just an API definition that is implemented in .cpp file.

Test Plan: At this moment any Java lambda that takes no arguments and returns nothing (void()) can be scheduled on the database thread and executed on the C++ side.

Reviewers: atul, jon, tomek

Reviewed By: tomek

Subscribers: ashoat, abosh

Differential Revision: https://phab.comm.dev/D5236