[web] Use <a> instead of <div> in some places
Summary:
[Part of ENG-843](https://linear.app/comm/issue/ENG-843)
According to the linked article that @tomek linked in the comments of the task, we should use `<a>` in some places (where it displays a new chat).

Test Plan: Check if all changed buttons look the same. Check if the changed css classes are used anywhere else. Check if rest of the app looks the same.

Reviewers: tomek, atul, abosh

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D5327