[native] Fix inline sidebar background color
Summary:
Fixes [[https://linear.app/comm/issue/ENG-1954/inlinesidebar-cuts-off-bottom-of-message-bubble-on-native | ENG-1954]].
It turns out that sidebar wrapper/holder (with 100% width) had background color set to black. And because it has border radius set, the top-left corner looked OK (or the top-right in case when viewer is author of the message).

Fixed it by removing the background color and making the view opaque.

Test Plan:
Tested in iOS simulator, used RN Element Inspector.

Temporarily set the background to green and disabled cornerRadius (1st screen) to show the element. The last screen shows final effect

{F194579}

{F194578}

{F194577}

Reviewers: ashoat, tomek, atul

Reviewed By: atul

Subscribers: atul, abosh

Differential Revision: https://phab.comm.dev/D5325