[services] set terraform provider arguments conditionally
Summary: we want to be able to use the same terraform configuration for both localstack and our production environment. in order to do this, we need to set the provider arguments (things like access_key) conditionally. the default will still be the localstack settings, but now we can override these by changing the terraform workspace to something else (e.g. "prod").

Test Plan: from the services directory, run `yarn init-local-cloud`.

Reviewers: atul, jon

Reviewed By: atul

Subscribers: ashoat, tomek, abosh

Differential Revision: https://phab.comm.dev/D5345