Add the match and parse functions for spoiler text
Summary: Added the match and parse functions in order to execute the RegEx
in the previous diff (in the stack), so we can then capture the spoiler
text and transform it later.

Test Plan: The testing for this will come in the subsequent diff, where we
will add the spoiler rule to the markdown rules in native.

Reviewers: atul!, ginsu, abosh

Subscribers: