[native] [9/34] ReactNav6: separate StackNavigationProp from StackNavigationHelpers
Summary:
We want to type `setParams` correctly for both the specific-screen and any-screen-on-a-navigator case. As a result, the type of `setParams` is incompatible between the two.

Note that I have left off `setParams` from `NavigationHelpers` in this diff. That's because I need to make some changes to our codebase before I can do that without leading to Flow errors... otherwise the two `setParams` will collide. I will reintroduce `setParams` to `NavigationHelpers` in a later diff in the stack.

Depends on D5401

Test Plan: Flow + ran the app on my phone and played around with it

Reviewers: atul, tomek, bartek

Subscribers: abosh

Differential Revision: https://phab.comm.dev/D5402