[services] Tunnelbroker - Update Tunnelbroker protobuf file with ping and new device token in bidirectional stream
Summary:
This diff introduces an update of Tunnelbroker protobuf bidirectional stream `MessagesStream` with the ping and device token fields.

In [[ https://linear.app/comm/issue/ENG-1766/make-sure-the-device-is-offline-to-send-a-notification-to-wake-it-up | ENG-1766 ]] we created pinging messages to make sure the client is online. Currently, we are using it in a unidirectional `Get`, but we should add it to the bidirectional `MessagesStream` as it will replace the `Send` and `Get` methods.

In [[ https://linear.app/comm/issue/ENG-1782/generate-notice-for-client-for-invalid-device-tokens-in-tunnelbroker | ENG-1782 ]] we've added the mechanism of requesting new device notification tokens. Currently, we are requesting it in the `Get` message and expecting it in `Send` from the client. These two fields must be merged into the bidirectional `MessagesStream`  because `Get` and `Send` will be deprecated in a favor of one bidirectional `MessagesStream`.

Related Linear task: [[ https://linear.app/comm/issue/ENG-2043/update-tunnelbroker-protobuf-bidirectional-stream-messagesstream-with | ENG-2043 ]]

Test Plan: Protobuf file compilation is successful.

Reviewers: jon, tomek, varun, ashoat

Reviewed By: jon, tomek, ashoat

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5361