[services] Tunnelbroker - Unifying paths in Docker according to the local repository
Summary:
This diff introduces changes to unify source paths according to the repository. These changes are required to remove the conditional paths includes in CMake and simplify future includes into the Cargo `build.rs` file as long as unifies the directory structure.

This diff also includes some minor prettifying to the `CMakeLists` and `Dockerfile` and removing unused commands from them.

Related Linear task: [[ https://linear.app/comm/issue/ENG-2044/fix-compiling-of-protobuf-file-in-cargos-buildrs-in-docker-for | ENG-2044 ]]

Test Plan: The service was successfully built in Docker and Nix dev environments.

Reviewers: jon, varun, marcin, tomek

Reviewed By: jon, tomek

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5380