[native] Display the spoiler content on tap
Summary:
We render a new Spoiler component for text content parsed to be spoilers. Upon tap, this Spoiler component reveals the text beneath
(https://linear.app/comm/issue/ENG-2011/display-the-spoiler-content-on-tap-within-messagelist-[native]). This diff will not be landed until ENG-2072 is addressed, to resolve the spoiler state not being remembered upon message tap.

Depends on D5373

Test Plan:
Spoiler messages are revealed when tapping the spoiler content.

{F202315}

Reviewers: atul, ginsu, abosh

Reviewed By: atul, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D5383