[native] added icons to tooltip item component
Summary:
added icons to the `TooltipItem` component. The five options of the `TooltipItem` are copy, reply, report, creae sidebar and open sidebar. Once we have the icons, we will convert the tooltip into the toolbar.

---

Linear Task: [[ https://linear.app/comm/issue/ENG-1541/move-message-tooltip-actions-to-the-bottom-to-make-space-for-reactions | ENG-1541 ]]

Test Plan:
Please review the screenshots to see the before and after for the changes i made:

Before:
{F201037}

After:
Create Thread Case:
{F201039}

Open Thread Case:
{F201042}

Reviewers: atul, rohan, abosh

Reviewed By: atul, rohan

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5374