[services] Tunnelbroker - Moving CMakeLists file to use for gTest only
Summary:
This diff is a **part of the flat diffs stack** to transform the current Tunnelbroker project into a Rust app by calling the current C++ codebase from Rust. The related transition task is [[ https://linear.app/comm/issue/ENG-1072 | ENG-1072 ]].

This diff moves the current `CMakeLists.txt` file into the gtest `test` directory and makes it reduced to build only the gtest app for unit testing purposes only.

Related Linear task and discussion: [[ https://linear.app/comm/issue/ENG-2083/changing-the-tunnelbroker-project-structure-to-be-a-rust-app#comment-c9252d58 | ENG-2083 ]]

**CI Notice:**
The CI will fail on this and further diffs until the D5461 where the changes to CI build commands are made.
This stack will be** landed all in one** to prevent CI from failing on diffs out of this stack.

Test Plan:
1. Test current diff CMakeLists test build:
Go to the C++ tests directory: `services/tunnelbroker/src/libcpp/test`.
Run the CMake build using Nix: `nix develop --accept-flake-config --command bash -c "cmake -B build . && make -C build -j"`.
The expected result is the successfully building of the `runTests` app for gTests.

2. Test completed stack:
To test it patch to D5436 using `arc patch D5436`, run `nix develop`.
Go to tunnelbroker service directory `services/tunnelbroker` and run the Cargo building process `cargo build` or `cargo run`.
The Rust application will be successfully built.

3. CI Testing:
These diff changes are included as a part of the stack in the later D5461 diff, where the changes to CI build commands are made. The CI succeded on the D5461 which reflects that this changes are passing the build.

Reviewers: jon, varun, marcin, atul

Reviewed By: jon, atul

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5430